lab 4: guild: enhances rfq:delivery_ready events with the shop lat/lng coordinates. driver: parses the lat/lng coordinates and stores them in the DeliveryRequest object, had to be stored as Doubles instead of Longs.

+++ driver/src/main/java/kpbinc/cs462/driver/model/DeliveryRequest.java
@@ -30,8 +30,8 @@ public class DeliveryRequest extends Delivery {
 	private Long shopId;
 	private String shopName;
 	private String shopAddress;
-	private Long shopLatitude;
-	private Long shopLongitude;
+	private Double shopLatitude;
+	private Double shopLongitude;
 	
 	private String estimatedDeliveryTime;
 	private Float bidAmount;
@@ -96,19 +96,19 @@ public class DeliveryRequest extends Delivery {
 		this.shopAddress = shopAddress;
 	}
 
-	public Long getShopLatitude() {
+	public Double getShopLatitude() {
 		return shopLatitude;
 	}
 
-	public void setShopLatitude(Long shopLatitude) {
+	public void setShopLatitude(Double shopLatitude) {
 		this.shopLatitude = shopLatitude;
 	}
 
-	public Long getShopLongitude() {
+	public Double getShopLongitude() {
 		return shopLongitude;
 	}
 
-	public void setShopLongitude(Long shopLongitude) {
+	public void setShopLongitude(Double shopLongitude) {
 		this.shopLongitude = shopLongitude;
 	}
